enables scf_type cholesky with polaritonic uccsd. also fixes yet more bugs in rcis code

+++ src/polaritonic_scf/uccsd.h
@@ -89,7 +89,7 @@ class PolaritonicUCCSD: public PolaritonicHF {
     std::shared_ptr<Matrix> oe_cavity_terms_;
 
     /// number of auxiliary basis functions
-    size_t nQ_;
+    long int nQ_;
 
     /// temporary container 1
     double * tmp1_;
